feature: RequestHelper
